Improved design of logger factory: made all sub classes private.
